refactor(archive): move archive helper to utils/ for #692 forward-compat
Relocate the generic download+extract helper from
src/apm_cli/marketplace/archive.py to src/apm_cli/utils/archive.py so a
future URL-package installer can reuse it without importing marketplace
internals. The module was already marketplace-free; this makes the
decoupling structural. Zip-slip / path-traversal / decompression-bomb
guards stay in the helper so they apply to every caller.

Honors #692 forward-compat constraint 3 (standalone archive extraction).
No production caller imported the helper yet; the only importer (the
safety test) moves to tests/unit/utils/test_archive.py.

Refs #692
